refactor #270: StoreMood 컨트롤러 분리
